Updated strings for file viewers. Redone file fragments structure.
