+++ app/src/main/res/layout/fragment_html_open.xml
@@ -9,7 +9,7 @@
         android:layout_height="wrap_content"
         android:gravity="center_horizontal"
         android:layout_centerInParent="true"
-        android:text="@string/fv_can_not_show_image_message"
+        android:text="@string/fv_can_not_show_message"
         android:textColor="@android:color/black"
         android:visibility="gone"/>
