+++ app/src/main/res/layout/fragment_image_open.xml
@@ -10,7 +10,7 @@
         android:layout_height="wrap_content"
         android:gravity="center_horizontal"
         android:layout_centerInParent="true"
-        android:text="@string/fv_can_not_show_html_message"
+        android:text="@string/fv_can_not_show_message"
         android:textColor="@android:color/black"
         android:visibility="gone"/>
